refactor(bundle): introduce Idle/Stored typestates
